pyproject.toml
[tool.poetry]
name = "sm-trendy"
version = "0.0.1"
description = "Retrieve google trends and store in the cloud"
authors = ["LM <hi@leima.is>"]
license = "GNU"
readme = "README.md"
packages = [{include = "sm_trendy"}]

[tool.poetry.dependencies]
python = "^3.10"
pytrends = "^4.9.2"
fastapi = "^0.100.0"
loguru = "^0.7.0"
pyarrow = "^12.0.1"
cloudpathlib = {version = "^0.15.1"}
s3fs = "^2023.6.0"
click = "^8.1.6"
python-slugify = "^8.0.1"
pydantic = "^2.0.3"
google-search-results = "^2.4.2"
python-dotenv = "^1.0.0"
fsspec = "^2023.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
trendy = "sm_trendy.cli:trendy"
